Implements a filesystem driver which provides the ability to mashup exisiting files in creative ways in-order to create new files.
Think of it as a way to union files; not filesystems.

Its a prototype idea with a prototype implementation.  Tested and
working on 3.0.9-rc1.  I have included Documentation file which details
the idea with examples and possible applications.

Any suggestions/ideas to make this useful and generally applicable is
very much appreciated!

Signed-off-by: Ram Pai <linuxram <at> us.ibm.com>