Uses extended attributes.
The file on which the mashed file is mounted, stores all the files
that the mashed file is made up off, along with its offset/len,
It stores all this information in its extended attributes.

Whenever a mount is attempted without any options, the options are read
in from the extended attributes of the mountpoint and mounted
accordingly.